+++ packages/common/src/utils/auth/createAuthUserInfo.ts
@@ -1,33 +1,6 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-import {get, has} from "lodash";
+import createAuthUser from "@sentrei/common/utils/auth/createAuthUser";
 
-export const createAuthUser = (
-  firebaseUser: firebase.User | null,
-): {
-  id: string;
-  email: string | null;
-  emailVerified: any;
-} | null => {
-  if (!firebaseUser || !firebaseUser.uid) {
-    return null;
-  }
-  return {
-    id: get(firebaseUser, "uid"),
-    email: get(firebaseUser, "email"),
-    emailVerified: has(firebaseUser, "emailVerified")
-      ? get(firebaseUser, "emailVerified") // client
-      : get(firebaseUser, "email_verified"), // admin
-  };
-};
-
-const createAuthUserInfo = ({firebaseUser = null, token = null} = {}): {
-  AuthUser: {
-    id: string;
-    email: string | null;
-    emailVerified: any;
-  } | null;
-  token: null;
-} => {
+const createAuthUserInfo = ({firebaseUser = null, token = null} = {}) => {
   return {
     AuthUser: createAuthUser(firebaseUser),
     token,