+++ packages/common/src/utils/auth/logout.ts
@@ -1,12 +1,12 @@
+/* globals window */
 import firebase from "firebase/app";
-
 import "firebase/auth";
-import isBrowser from "@sentrei/common/utils/isBrowser";
 
-export default async function logout(): Promise<boolean> {
+export default async () => {
   try {
     await firebase.auth().signOut();
-    if (isBrowser()) {
+    // Sign-out successful.
+    if (typeof window !== "undefined") {
       // Remove the server-side rendered user data element. See:
       // https://github.com/zeit/next.js/issues/2252#issuecomment-353992669
       const elem = window.document.getElementById("__MY_AUTH_USER_INFO");
@@ -20,4 +20,4 @@ export default async function logout(): Promise<boolean> {
     console.error(e);
     return false;
   }
-}
+};