+++ packages/ui/src/components/Auth/Auth.tsx
@@ -90,13 +90,10 @@ export default function Auth({type}: Props): JSX.Element {
           firebase.auth().sendPasswordResetEmail(data.email);
           handleSuccess("Please check your email");
         } catch (err) {
-          // TODO: Temporary for now
-          // eslint-disable-next-line no-console
-          console.log(err);
           handleError(err);
         }
         break;
-      case authType.signin:
+      case authType.login:
         try {
           await firebase
             .auth()
@@ -104,9 +101,6 @@ export default function Auth({type}: Props): JSX.Element {
           setOpen(false);
           Router.push("/");
         } catch (err) {
-          // TODO: Temporary for now
-          // eslint-disable-next-line no-console
-          console.log(err);
           handleError(err);
         }
         break;
@@ -118,9 +112,6 @@ export default function Auth({type}: Props): JSX.Element {
           setOpen(false);
           Router.push("/");
         } catch (err) {
-          // TODO: Temporary for now
-          // eslint-disable-next-line no-console
-          console.log(err);
           handleError(err);
         }
         break;
@@ -139,12 +130,12 @@ export default function Auth({type}: Props): JSX.Element {
       <div className={classes.paper}>
         <Avatar className={classes.avatar}>
           {type === authType.reset ? <MailOutlinedIcon /> : null}
-          {type === authType.signin ? <LockOutlinedIcon /> : null}
+          {type === authType.login ? <LockOutlinedIcon /> : null}
           {type === authType.signup ? <AccountCircleOutlinedIcon /> : null}
         </Avatar>
         <Typography component="h1" variant="h3">
           {type === authType.reset ? "Reset email" : null}
-          {type === authType.signin ? "Sign in" : null}
+          {type === authType.login ? "Log in" : null}
           {type === authType.signup ? "Sign up" : null}
         </Typography>
         <form
@@ -175,7 +166,7 @@ export default function Auth({type}: Props): JSX.Element {
             control={control}
             defaultValue=""
           />
-          {type === authType.signin || type === authType.signup ? (
+          {type === authType.login || type === authType.signup ? (
             <Controller
               as={
                 <TextField
@@ -198,7 +189,7 @@ export default function Auth({type}: Props): JSX.Element {
               defaultValue=""
             />
           ) : null}
-          {type === authType.signin || type === authType.signup ? (
+          {type === authType.login || type === authType.signup ? (
             <FormControlLabel
               control={<Checkbox value="remember" color="primary" />}
               label="Remember me"
@@ -212,14 +203,14 @@ export default function Auth({type}: Props): JSX.Element {
             className={classes.submit}
           >
             {type === authType.reset ? "Send reset email" : null}
-            {type === authType.signin ? "Sign in" : null}
+            {type === authType.login ? "Log in" : null}
             {type === authType.signup ? "Sign up" : null}
           </Button>
         </form>
-        {type === authType.signin ? (
+        {type === authType.login ? (
           <Grid container>
             <Grid item xs>
-              <Link href="/reset" variant="body2">
+              <Link href="/reset-password" variant="body2">
                 Forgot password?
               </Link>
             </Grid>
@@ -233,8 +224,8 @@ export default function Auth({type}: Props): JSX.Element {
         {type === authType.signup ? (
           <Grid container justify="center">
             <Grid item>
-              <Link href="/signin" variant="body2">
-                Already have an account? Sign in
+              <Link href="/login" variant="body2">
+                Already have an account? Log in
               </Link>
             </Grid>
           </Grid>