+++ packages/ui/src/hooks/useFirebaseAuth/useFirebaseAuth.ts
@@ -1,10 +1,10 @@
 import firebase from "firebase/app";
 import React from "react";
 
-const useAuth = (): {
-  initializing: boolean;
-  user: firebase.User | null;
-} => {
+import "firebase/auth";
+import setSession from "@sentrei/common/utils/firebaseSessionHandler";
+
+const useFirebaseAuth = () => {
   const [state, setState] = React.useState(() => {
     const user = firebase.auth().currentUser;
     return {
@@ -13,16 +13,22 @@ const useAuth = (): {
     };
   });
 
-  function onChange(user: firebase.User | null): void {
+  function onChange(user: firebase.User | null) {
     setState({initializing: false, user});
+
+    // Call server to update session.
+    setSession(user);
   }
 
   React.useEffect(() => {
+    // Listen for auth state changes.
     const unsubscribe = firebase.auth().onAuthStateChanged(onChange);
-    return (): void => unsubscribe();
+
+    // Unsubscribe to the listener when unmounting.
+    return () => unsubscribe();
   }, []);
 
   return state;
 };
 
-export default useAuth;
+export default useFirebaseAuth;