+++ packages/web/cypress/plugins/cy-ts-preprocessor.js
@@ -1,6 +1,3 @@
-/* eslint-disable import/no-extraneous-dependencies */
-/* eslint-disable import/no-unresolved */
-/* eslint-disable @typescript-eslint/no-var-requires */
 const wp = require("@cypress/webpack-preprocessor");
 
 const webpackOptions = {