+++ packages/web/cypress/plugins/index.js
@@ -1,11 +1,7 @@
-/* eslint-disable import/no-extraneous-dependencies */
-/* eslint-disable global-require */
-/* eslint-disable @typescript-eslint/no-var-requires */
-/* eslint-disable @typescript-eslint/explicit-function-return-type */
+const cypressTypeScriptPreprocessor = require("./cy-ts-preprocessor");
 
-module.exports = (on, config) => {
-  require("@cypress/code-coverage/task")(on, config);
-  on("file:preprocessor", require("./cy-ts-preprocessor"));
-  on("file:preprocessor", require("@cypress/code-coverage/use-babelrc"));
-  return config;
+module.exports = on => {
+  // require("@cypress/code-coverage/task")(on, config);
+  on("file:preprocessor", cypressTypeScriptPreprocessor);
+  // on("file:preprocessor", require("@cypress/code-coverage/use-babelrc"));
 };