+++ packages/web/src/middleware/cookieSession.ts
@@ -3,10 +3,21 @@ import cookieSession from "cookie-session";
 import {NextApiRequest, NextApiResponse} from "next";
 import getConfig from "next/config";
 
+import isDev from "@sentrei/common/utils/isDev";
+
 const {publicRuntimeConfig} = getConfig();
-export const addSession = (req: any, res: any): any => {
-  // Ensure that session secrets are set.
-  if (
+export const addSession = (req: any, res: any) => {
+  // An array is useful for rotating secrets without invalidating old sessions.
+  // The first will be used to sign cookies, and the rest to validate them.
+  // https://github.com/expressjs/cookie-session#keys
+  let sessionSecrets = <any>[];
+
+  if (isDev()) {
+    sessionSecrets = [
+      process.env.SESSION_SECRET_CURRENT,
+      process.env.SESSION_SECRET_PREVIOUS,
+    ];
+  } else if (
     !(
       publicRuntimeConfig.SESSION_SECRET_CURRENT &&
       publicRuntimeConfig.SESSION_SECRET_PREVIOUS
@@ -15,15 +26,12 @@ export const addSession = (req: any, res: any): any => {
     throw new Error(
       "Session secrets must be set as env vars `SESSION_SECRET_CURRENT` and `SESSION_SECRET_PREVIOUS`.",
     );
+  } else {
+    sessionSecrets = [
+      publicRuntimeConfig.SESSION_SECRET_CURRENT,
+      publicRuntimeConfig.SESSION_SECRET_PREVIOUS,
+    ];
   }
-
-  // An array is useful for rotating secrets without invalidating old sessions.
-  // The first will be used to sign cookies, and the rest to validate them.
-  // https://github.com/expressjs/cookie-session#keys
-  const sessionSecrets = [
-    publicRuntimeConfig.SESSION_SECRET_CURRENT,
-    publicRuntimeConfig.SESSION_SECRET_PREVIOUS,
-  ];
 
   // Example:
   // https://github.com/billymoon/micro-cookie-session
@@ -41,7 +49,7 @@ export const addSession = (req: any, res: any): any => {
 export default (handler: any) => (
   req: NextApiRequest,
   res: NextApiResponse,
-): any => {
+) => {
   try {
     addSession(req, res);
   } catch (e) {