+++ packages/web/src/pages/about.tsx
@@ -22,7 +22,7 @@ function About({t}: any): JSX.Element {
         featuresText={t("features")}
         pricingText={t("pricing")}
         productText={t("product")}
-        signInText={t("signIn")}
+        logInText={t("logIn")}
         signUpText={t("signUp")}
         testimonialText={t("testimonial")}
       />